Fix circular dependency by moving waitWhileLoading to waits-utils
