+++ e2e/app/component/concept-domain-card.ts
@@ -1,8 +1,8 @@
 import {Page} from 'puppeteer';
-import {waitWhileLoading} from 'utils/test-utils';
 import Button from 'app/element/button';
 import Container from 'app/container';
 import {getPropValue} from 'utils/element-utils';
+import {waitWhileLoading} from '../../utils/waits-utils';
 
 export enum Domain {
    Conditions = 'Conditions',
