+++ e2e/app/page/cohort-build-page.ts
@@ -5,8 +5,7 @@ import Button from 'app/element/button';
 import ClrIconLink from 'app/element/clr-icon-link';
 import {ElementType} from 'app/xpath-options';
 import {makeRandomName} from 'utils/str-utils';
-import {waitWhileLoading} from 'utils/test-utils';
-import {waitForDocumentTitle, waitForNumericalString} from 'utils/waits-utils';
+import {waitForDocumentTitle, waitForNumericalString, waitWhileLoading} from 'utils/waits-utils';
 import {buildXPath} from 'app/xpath-builders';
 import {LinkText} from 'app/text-labels';
 import AuthenticatedPage from './authenticated-page';
