+++ e2e/app/page/cohort-participant-detail-page.ts
@@ -1,6 +1,5 @@
 import {Page} from 'puppeteer';
-import {waitWhileLoading} from 'utils/test-utils';
-import {waitForDocumentTitle} from 'utils/waits-utils';
+import {waitForDocumentTitle, waitWhileLoading} from 'utils/waits-utils';
 import Button from 'app/element/button';
 import {LinkText} from 'app/text-labels';
 import AuthenticatedPage from './authenticated-page';
