+++ e2e/app/page/cohort-search-page.ts
@@ -5,8 +5,8 @@ import Table from 'app/component/table';
 import Button from 'app/element/button';
 import ClrIconLink from 'app/element/clr-icon-link';
 import Textbox from 'app/element/textbox';
-import {centerPoint, dragDrop, waitWhileLoading} from 'utils/test-utils';
-import {waitForNumericalString} from 'utils/waits-utils';
+import {centerPoint, dragDrop} from 'utils/test-utils';
+import {waitForNumericalString, waitWhileLoading} from 'utils/waits-utils';
 import {LinkText} from 'app/text-labels';
 import {waitUntilChanged} from 'utils/element-utils';
 import AuthenticatedPage from './authenticated-page';
