+++ e2e/app/page/conceptset-actions-page.ts
@@ -1,6 +1,5 @@
 import {Page} from 'puppeteer';
-import {waitWhileLoading} from 'utils/test-utils';
-import {waitForDocumentTitle} from 'utils/waits-utils';
+import {waitForDocumentTitle, waitWhileLoading} from 'utils/waits-utils';
 import Button from 'app/element/button';
 import {LinkText} from 'app/text-labels';
 import ConceptDomainCard, {Domain} from 'app/component/concept-domain-card';
