+++ e2e/app/page/dataset-build-page.ts
@@ -2,8 +2,7 @@ import Table from 'app/component/table';
 import Button from 'app/element/button';
 import Checkbox from 'app/element/checkbox';
 import ClrIconLink from 'app/element/clr-icon-link';
-import {waitWhileLoading} from 'utils/test-utils';
-import {waitForDocumentTitle} from 'utils/waits-utils';
+import {waitForDocumentTitle, waitWhileLoading} from 'utils/waits-utils';
 import {buildXPath} from 'app/xpath-builders';
 import {ElementType} from 'app/xpath-options';
 import AuthenticatedPage from './authenticated-page';
