+++ e2e/app/page/workspace-edit-page.ts
@@ -6,8 +6,7 @@ import Textbox from 'app/element/textbox';
 import WebComponent from 'app/element/web-component';
 import {ElementType} from 'app/xpath-options';
 import {ElementHandle, Page} from 'puppeteer';
-import {waitWhileLoading} from 'utils/test-utils';
-import {waitForDocumentTitle} from 'utils/waits-utils';
+import {waitForDocumentTitle, waitWhileLoading} from 'utils/waits-utils';
 import {buildXPath} from 'app/xpath-builders';
 import {LinkText} from 'app/text-labels';
 import WorkspaceBase, {UseFreeCredits} from './workspace-base';
