+++ e2e/app/page/workspaces-page.ts
@@ -4,8 +4,8 @@ import {Language, LinkText, PageUrl} from 'app/text-labels';
 import WorkspaceEditPage, {FIELD as EDIT_FIELD} from 'app/page/workspace-edit-page';
 import {makeWorkspaceName} from 'utils/str-utils';
 import RadioButton from 'app/element/radiobutton';
-import {findWorkspace, waitWhileLoading} from 'utils/test-utils';
-import {waitForDocumentTitle, waitForText} from 'utils/waits-utils';
+import {findWorkspace} from 'utils/test-utils';
+import {waitForDocumentTitle, waitForText, waitWhileLoading} from 'utils/waits-utils';
 import ReactSelect from 'app/element/react-select';
 import WorkspaceDataPage from './workspace-data-page';
 import WorkspaceAnalysisPage from './workspace-analysis-page';
