+++ e2e/tests/datasets/dataset-create.spec.ts
@@ -3,8 +3,8 @@ import ClrIconLink from 'app/element/clr-icon-link';
 import CohortBuildPage from 'app/page/cohort-build-page';
 import WorkspaceDataPage from 'app/page/workspace-data-page';
 import {Option, ResourceCard} from 'app/text-labels';
-import {findWorkspace, signIn, waitWhileLoading} from 'utils/test-utils';
-import {waitForText} from 'utils/waits-utils';
+import {findWorkspace, signIn} from 'utils/test-utils';
+import {waitForText, waitWhileLoading} from 'utils/waits-utils';
 
 describe('Dataset test', () => {
