+++ e2e/tests/notebook/reader-copy-to-workspace.spec.ts
@@ -8,7 +8,8 @@ import WorkspaceDataPage from 'app/page/workspace-data-page';
 import {Option, Language, LinkText, ResourceCard, WorkspaceAccessLevel} from 'app/text-labels';
 import {config} from 'resources/workbench-config';
 import {makeRandomName} from 'utils/str-utils';
-import {findWorkspace, signIn, signInAs, signOut, waitWhileLoading} from 'utils/test-utils';
+import {findWorkspace, signIn, signInAs, signOut} from 'utils/test-utils';
+import {waitWhileLoading} from '../../utils/waits-utils';
 
 jest.setTimeout(20 * 60 * 1000);
