+++ e2e/tests/workspace/workspace-share.spec.ts
@@ -6,8 +6,9 @@ import WorkspaceAboutPage from 'app/page/workspace-about-page';
 import WorkspacesPage from 'app/page/workspaces-page';
 import {Option, LinkText, WorkspaceAccessLevel} from 'app/text-labels';
 import {config} from 'resources/workbench-config';
-import {findWorkspace, signIn, signInAs, signOut, waitWhileLoading} from 'utils/test-utils';
+import {findWorkspace, signIn, signInAs, signOut} from 'utils/test-utils';
 import WorkspaceDataPage from 'app/page/workspace-data-page';
+import {waitWhileLoading} from '../../utils/waits-utils';
 
 describe('Share workspace', () => {
