+++ e2e/utils/test-utils.ts
@@ -5,10 +5,10 @@ import Textarea from 'app/element/textarea';
 import Textbox from 'app/element/textbox';
 import GoogleLoginPage from 'app/page/google-login';
 import HomePage, {LabelAlias} from 'app/page/home-page';
-import {XPathOptions, ElementType} from 'app/xpath-options';
+import {ElementType, XPathOptions} from 'app/xpath-options';
 import * as fp from 'lodash/fp';
 import {ElementHandle, Page} from 'puppeteer';
-import {waitForText} from 'utils/waits-utils';
+import {waitForText, waitWhileLoading} from 'utils/waits-utils';
 import WorkspaceCard from 'app/component/workspace-card';
 import {WorkspaceAccessLevel} from 'app/text-labels';
 import WorkspacesPage from 'app/page/workspaces-page';
@@ -49,29 +49,6 @@ export async function signInAs(page: Page, userId: string, passwd: string, opts:
 export async function signOut(page: Page) {
   await Navigation.navMenu(page, NavLink.SIGN_OUT);
   await page.waitFor(1000);
-}
-
-/**
- * <pre>
- * Wait while the page is loading (spinner is spinning and visible). Waiting stops when spinner stops spinning or when timed out.
- * It usually indicates the page is ready for user interaction.
- * </pre>
- */
-export async function waitWhileLoading(page: Page, timeOut: number = 90000): Promise<void> {
-  const notBlankPageSelector = '[data-test-id="sign-in-container"], title:not(empty), div.spinner, svg[viewBox]';
-  const spinElementsSelector = '[style*="running spin"], .spinner:empty, [style*="running rotation"]';
-
-  await Promise.race([
-    // To prevent checking on blank page, wait for elements exist in DOM.
-    page.waitForSelector(notBlankPageSelector),
-    page.waitForSelector(spinElementsSelector),
-  ]);
-
-  // Wait for spinners stop and gone.
-  await page.waitForFunction((css) => {
-    const elements = document.querySelectorAll(css);
-    return elements && elements.length === 0;
-  }, {polling: 'mutation', timeout: timeOut}, spinElementsSelector);
 }
